+++ idp-server/src/test/java/de/gematik/idp/token/AccessTokenBuilderTest.java
@@ -17,6 +17,7 @@
 package de.gematik.idp.token;
 
 import static de.gematik.idp.IdpConstants.EREZEPT;
+import static de.gematik.idp.IdpConstants.OID_VERSICHERTER;
 import static de.gematik.idp.IdpConstants.OPENID;
 import static de.gematik.idp.IdpConstants.PAIRING;
 import static de.gematik.idp.brainPoolExtension.BrainpoolAlgorithmSuiteIdentifiers.BRAINPOOL256_USING_SHA256;
@@ -24,6 +25,7 @@ import static de.gematik.idp.field.ClaimName.ALGORITHM;
 import static de.gematik.idp.field.ClaimName.AUDIENCE;
 import static de.gematik.idp.field.ClaimName.AUTH_TIME;
 import static de.gematik.idp.field.ClaimName.CLIENT_ID;
+import static de.gematik.idp.field.ClaimName.DISPLAY_NAME;
 import static de.gematik.idp.field.ClaimName.FAMILY_NAME;
 import static de.gematik.idp.field.ClaimName.GIVEN_NAME;
 import static de.gematik.idp.field.ClaimName.ID_NUMBER;
@@ -66,6 +68,12 @@ class AccessTokenBuilderTest {
   private static final String KEY_ID = "my_key_id";
   private static final String EREZEPT_AUDIENCE = "erezeptAudience";
   private static final String PAIRING_AUDIENCE = "pairingAudience";
+
+  private static final String AUTH_CODE_SEKTORALER_IDP =
+      "eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6InB1a19mZF9zaWcifQ.eyJwcm9mZXNzaW9uT0lEIjoiMS4yLjI3Ni4wLjc2LjQuNDkiLCJvcmdhbml6YXRpb25OYW1lIjoiMTA5NTAwOTY5IiwiaWROdW1tZXIiOiJYMTEwNDExNjc1IiwiYW1yIjpbIm1mYSJdLCJpc3MiOiJodHRwczovL2lkcGZhZGkuZGV2LmdlbWF0aWsuc29sdXRpb25zIiwicmVzcG9uc2VfdHlwZSI6bnVsbCwic25jIjoiS1RtWHB6aURXaFBIUjBHS2RUdEY5eV9QeHN5NTRWbWwiLCJjb2RlX2NoYWxsZW5nZV9tZXRob2QiOiJmcm9udGVuZENvZGVDaGFsbGVuZ2VNZXRob2QiLCJkaXNwbGF5X25hbWUiOiJEYXJpdXMgTWljaGFlbCBCcmlhbiBVYmJvIEdyYWYgdm9uIELDtmRlZmVsZCIsInRva2VuX3R5cGUiOiJjb2RlIiwiY2xpZW50X2lkIjoiZnJvbnRlbmRDbGllbnRJZCIsInNjb3BlIjoib3BlbmlkIGUtcmV6ZXB0IiwiYXV0aF90aW1lIjoxNjkwMjAyMzI1LCJyZWRpcmVjdF91cmkiOiJmcm9udGVuZFJlZGlyZWN0VXJpIiwic3RhdGUiOiJmcm9udGVuZFN0YXRlIiwiZXhwIjoxNjkwMjA1OTQyLCJpYXQiOjE2OTAyMDIzMjUsImNvZGVfY2hhbGxlbmdlIjoiZnJvbnRlbmRDb2RlQ2hhbGxlbmdlIiwianRpIjoiOGIyOWI0NTlkMTNiY2NmZCJ9.UIObtP1dS3iuXBkUOgPaAoQm1kHlbe6HRWeweP8-CeizPPQmxjQzHSbevm7o8-UzkllUaSyndliIyeTdV6wJdA";
+
+  private static final String AUTH_CODE_FOR_SMCB =
+      "eyJhbGciOiJCUDI1NlIxIiwidHlwIjoiSldUIiwia2lkIjoicHVrX2lkcF9zaWcifQ.eyJvcmdhbml6YXRpb25OYW1lIjoiQXNjaG9mZnNjaGUgQXBvdGhla2UgVEVTVC1PTkxZIiwicHJvZmVzc2lvbk9JRCI6IjEuMi4yNzYuMC43Ni40LjU0IiwiaWROdW1tZXIiOiIzLTItRVBBLTgzMzYyMTk5OTc0MTYwMCIsImFtciI6WyJtZmEiLCJzYyIsInBpbiJdLCJpc3MiOiJodHRwczovL2lkcC56ZW50cmFsLmlkcC5zcGxpdGRucy50aS1kaWVuc3RlLmRlIiwicmVzcG9uc2VfdHlwZSI6ImNvZGUiLCJzbmMiOiJRa1hSVmZvNTNFZ0IybFhCcFp0b0kyTTZjMGE1bFg2TiIsImNvZGVfY2hhbGxlbmdlX21ldGhvZCI6IlMyNTYiLCJnaXZlbl9uYW1lIjpudWxsLCJ0b2tlbl90eXBlIjoiY29kZSIsIm5vbmNlIjoiOHVRaURrODBTMjl6QTNWa1lzWjJWZVV0R1N1d2lOcDgiLCJjbGllbnRfaWQiOiJlUmV6ZXB0QXBwIiwic2NvcGUiOiJlLXJlemVwdCBvcGVuaWQiLCJhdXRoX3RpbWUiOjE2OTA0MzgxMTAsInJlZGlyZWN0X3VyaSI6Imh0dHA6Ly9yZWRpcmVjdC5nZW1hdGlrLmRlL2VyZXplcHQiLCJzdGF0ZSI6IndCOWE0cUZZRk1LeVJoRW9pdkRTbDN5eUFaWF9ua3o5IiwiZXhwIjoxNjkwNDM4MTcwLCJmYW1pbHlfbmFtZSI6bnVsbCwiaWF0IjoxNjkwNDM4MTEwLCJjb2RlX2NoYWxsZW5nZSI6Inh3S2t2dlZpU293LTFybnQxLXBMd3RMVFBXQWxzaDdYM1FNd0ZFZDlIWDgiLCJqdGkiOiIzMjc3NTJkOTNlZWE4NTkzIn0.AAOZW9pywmyY3xrHOjnBAZF1OhmiuFQmSRNwvMIDbaGi2bL1HIX943fT8MaXFOFI-q2_kJfj0hpgMOK1omUdvw";
 
   static {
     Security.removeProvider(BouncyCastleProvider.PROVIDER_NAME);
@@ -111,7 +119,7 @@ class AccessTokenBuilderTest {
             CLIENT_ID.getJoseName(),
             TestConstants.CLIENT_ID_E_REZEPT_APP,
             SCOPE.getJoseName(),
-            EREZEPT));
+            EREZEPT + " " + OPENID));
   }
 
   private void createAuthenticationTokenByBodyClaims(final Map<String, Object> map) {
@@ -186,7 +194,7 @@ class AccessTokenBuilderTest {
   }
 
   @Test
-  void verifyAudienceByScopeERezept() {
+  void verifyAudienceAndDisplayNameByScopeERezept() {
     createAuthenticationTokenByBodyClaims(
         Map.of(
             CLIENT_ID.getJoseName(),
@@ -195,10 +203,11 @@ class AccessTokenBuilderTest {
             OPENID + " " + EREZEPT));
     final JsonWebToken accessToken = accessTokenBuilder.buildAccessToken(authenticationToken);
     assertThat(accessToken.getBodyClaim(AUDIENCE)).get().isEqualTo(EREZEPT_AUDIENCE);
+    assertThat(accessToken.getBodyClaim(DISPLAY_NAME)).get().isEqualTo("Juna Fuchs");
   }
 
   @Test
-  void verifyAudienceByScopePairing() {
+  void verifyAudienceAndDisplayNameByScopePairing() {
     createAuthenticationTokenByBodyClaims(
         Map.of(
             CLIENT_ID.getJoseName(),
@@ -207,5 +216,31 @@ class AccessTokenBuilderTest {
             OPENID + " " + PAIRING));
     final JsonWebToken accessToken = accessTokenBuilder.buildAccessToken(authenticationToken);
     assertThat(accessToken.getBodyClaim(AUDIENCE)).get().isEqualTo(PAIRING_AUDIENCE);
+    assertThat(accessToken.getBodyClaims()).doesNotContainKey(DISPLAY_NAME.getJoseName());
+  }
+
+  @Test
+  void verifyNoDisplayNameForAnSmcbWithNoDisplayNameInAuthCode() {
+    final JsonWebToken authCode = new JsonWebToken(AUTH_CODE_FOR_SMCB);
+    assertThat(authCode).isNotNull();
+    final JsonWebToken accessToken = accessTokenBuilder.buildAccessToken(authCode);
+    assertThat(accessToken.getBodyClaims()).doesNotContainKey(DISPLAY_NAME.getJoseName());
+  }
+
+  @Test
+  void buildAccessTokenFromSektoralIdpAuthCode() {
+    final JsonWebToken authCode = new JsonWebToken(AUTH_CODE_SEKTORALER_IDP);
+    assertThat(authCode).isNotNull();
+    final JsonWebToken accessToken = accessTokenBuilder.buildAccessToken(authCode);
+    assertThat(accessToken.getBodyClaim(DISPLAY_NAME))
+        .get()
+        .isEqualTo(authCode.getBodyClaim(DISPLAY_NAME).orElseThrow());
+    assertThat(accessToken.getBodyClaim(ORGANIZATION_NAME))
+        .get()
+        .isEqualTo(authCode.getBodyClaim(ORGANIZATION_NAME).orElseThrow());
+    assertThat(accessToken.getBodyClaim(PROFESSION_OID)).get().isEqualTo(OID_VERSICHERTER);
+    assertThat(accessToken.getBodyClaim(ID_NUMBER))
+        .get()
+        .isEqualTo(authCode.getBodyClaim(ID_NUMBER).orElseThrow());
   }
 }